#730 move pidgin config functions to pidgin_config.py
