+++ src/a16_pidgin_logic/pidgin_config.py
@@ -162,3 +162,29 @@ def get_quick_pidgens_column_ref() -> dict[str, set[str]]:
             "otx_way",
         },
     }
+
+
+def pidginable_class_types() -> set:
+    return {"NameUnit", "LabelUnit", "TagUnit", "WayUnit"}
+
+
+def pidginable_atom_args() -> set:
+    return {
+        "acct_name",
+        "awardee_label",
+        "context",
+        "face_name",
+        "fcontext",
+        "fisc_tag",
+        "fneed",
+        "group_label",
+        "healer_name",
+        "hour_tag",
+        "idea_way",
+        "month_tag",
+        "need",
+        "owner_name",
+        "team_label",
+        "timeline_tag",
+        "weekday_tag",
+    }
